Add color and hovering effect
Add a colored box format to each div by selecting the gridPixel class in EAS.css and setting the appropriate attributes. Add a container div in index.html body that contains the grid. Add hovering effect with DOM manipulation and event handling in EAS.js , changing the div color when the mouse hovers a div inside the grid.