Add settings menu
Add divs to index.html with settings buttons. Turn the grid into a function to implement user input reaction properly in EAS.js and make clear button clean the grid on click by turning the grid's divs background color back to it's original color. Style the position of the grid and the menu in the page as the page's colors in EAS.css.